requirements.txt
eyed3
python-telegram-bot
gunicorn==20.1.0